+++ backend/database/database_config.py
@@ -1,4 +1,4 @@
 #type="production"
 type="local"
-#jwt_key_type = "testing"
-jwt_key_type = "production"
+jwt_key_type = "testing"
+#jwt_key_type = "production"